fixing issues after kraan's review, second round
